06-배진우/2024_겨울특강/과제/code/requirements.txt
TextLoader
DirectoryLoader
PyPDFLoader
UnstructuredCSVLoader
xml.etree.ElementTree
pandas
os
bs4
langchain-community
re
shutil
langchain_openai
python-dotenv
llama-index-readers-file
llama-parse
llama-index-core
langchain-teddynote
langchain-upstage
langchain-teddynote